+++ src/index.ts
@@ -7,7 +7,6 @@ import * as redis from 'redis';
 import * as AdminService from './service/admin';
 import * as AggregateService from './service/aggregate';
 import * as NotificationService from './service/notification';
-import * as OrderService from './service/order';
 import * as PerformanceService from './service/performance';
 import * as ReserveService from './service/reserve';
 import * as TaskService from './service/task';
@@ -29,7 +28,6 @@ export import AWS = AWS;
 export namespace service {
     export import admin = AdminService;
     export import notification = NotificationService;
-    export import order = OrderService;
     export import performance = PerformanceService;
     export import reserve = ReserveService;
     export import aggregate = AggregateService;